requirements.txt
Flask==2.0.3
mysql_connector_repackaged==0.3.1
pandas==1.4.2